feat: 拆分quickstart为consumer和provider 2个工程
