+++ spring-cloud-tencent-examples/polaris-quickstart-example/quickstart-consumer/src/main/java/com/tencent/cloud/polaris/quickstart/consumer/EchoServer.java
@@ -15,19 +15,15 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.tencent.cloud.polaris.quickstart.example;
+package com.tencent.cloud.polaris.quickstart.consumer;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
-@SpringBootApplication
-@EnableDiscoveryClient
-@EnableFeignClients
-public class EchoServiceApplication {
+@FeignClient(value = "EchoServer")
+public interface EchoServer {
 
-    public static void main(String[] args) {
-        SpringApplication.run(EchoServiceApplication.class, args);
-    }
+    @GetMapping("/quickstart/echo")
+    String echo(@RequestParam("value") String value);
 }
