+++ spring-cloud-tencent-examples/polaris-quickstart-example/quickstart-consumer/src/main/resources/log4j.properties
